+++ ada-malef/src-base/malef-wrapper.ads
@@ -26,15 +26,29 @@
 --                                                                           --
 -------------------------------------------------------------------------------
 
---
--- @summary
---
---
--- @description
---
-package Malef.Wrapper is
 
-end Malef.Wrapper;
+--
+-- This function wraps another function defined by the user in order to restore
+-- the terminal in case the user hasn't done it. That way we can be sure the
+-- terminal isn't damaged after execution.
+--
+-- @param Parameters
+-- The parameters the user has to enter (it's a record type).
+--
+-- @param User_Function
+-- The user defined function which is executed inside the wrapper.
+--
+-- @return
+-- It returns another record type.
+--
+generic
+   type Parameters_Type (<>) is private;
+   type Return_Type is private;
+   type User_Function_Type is access function (Parameters : Parameters_Type)
+                                               return Return_Type;
+function Malef.Wrapper (Parameters    : Parameters_Type;
+                        User_Function : User_Function_Type)
+                  return Return_Type;
 
 ---=======================-------------------------=========================---
 --=======================-- E N D   O F   F I L E --=========================--