+++ dev-tools/do.py
@@ -40,8 +40,16 @@ import os
 import sys
 
 
-dirs = ["ada-malef/src", "ada-malef/src/linux", "ada-malef/src/windows"]
+dirs = ["ada-malef"]
+temp = dirs.copy()
 
+while len(temp) != 0:
+    d = temp.pop(0)
+    _ = filter(os.path.isdir, [os.path.join(d, f) for f in os.listdir(d)])
+    dirs += _
+    temp += _
+
+del temp
 
 def _todo():
     TODO = {}
@@ -76,14 +84,75 @@ def _todo():
                         (l, data.lstrip(' ')))
 
 
+def _commit():
+    def read (filename):
+        file = open(filename, 'r')
+        data = file.read()
+        file.close()
+        del file;
+        return data;
+
+    def status (filename, status, description=""):
+        print("\033[36;1m%s\033[0m: " % filename, end="")
+        if status:
+            print("\033[32;1m[DONE]\033[0m")
+        else:
+            print("\033[31;1m[FAIL]\033[0m \033[2m%s\033[0m" % description)
+
+    version = read("VERSION").replace('\n', "")
+    done = True
+
+    # alire.toml
+    alire_toml = read("alire.toml").split('\n')
+    for line in alire_toml:
+        if '=' in line:
+            line = line.split(" = ")
+            if line[0] == "version":
+                if line[1] != '"%s"' % version:
+                    status("alire.toml", False,
+                            'Change: version = "%s"' % version)
+                    done = False
+                else:
+                    status("alire.toml", True)
+                break;
+
+
+    # CHANGELOG.md
+    changelog_md = read("CHANGELOG.md").split('\n')
+    for line in changelog_md:
+        if line[:3] == "## ":
+            line = line.split(' ')[1]
+            if len(line.split('.')) != 3:
+                continue;
+            if line != "[%s]" % version:
+                status("CHANGELOG.md", False, "Add new version [%s]" % version)
+                done = False
+            else:
+                status("CHANGELOG.md", True)
+            break;
+
+    if not done:
+        print("\033[31mYou can't commit just yet!\033[0m")
+    else:
+        print("You can commit!")
+        print("Remember to check:")
+        for t in ["It compiles for all systems",
+                  "You've added a description"]:
+            print(" \033[33m+\033[0m %s" % t)
+
+
 commands = {
         "linux": "gprbuild -p -Pmalef",
         "windows": "wine gprbuild -p -Pmalef "
                    "-XMALEF_OPERATING_SYSTEM=windows",
         "tests-linux": "gprbuild -p -Ptests/ada/tests.gpr",
-        "tests-windows": "wine grpbuild -p -Ptests/ada/tests.gpr",
+        "tests-windows": ["wine gprbuild -p -Ptests/ada/tests.gpr "
+                          "-XMALEF_OPERATING_SYSTEM=windows",
+                          "cp alire/build/lib-windows/libMalef.dll "
+                          "tests/ada/bin/"],
         "docs": "gnatdoc -bplwPmalef --enable-build",
-        "todo": _todo
+        "todo": _todo,
+        "commit": _commit,
 }
 
 
@@ -107,10 +176,12 @@ def main():
     cmd = commands[sys.argv[1]]
 
     if isinstance(cmd, str):
+        print("\033[2m%s\033[0m" % cmd)
         os.system(cmd)
     elif isinstance(cmd, list):
         for c in cmd:
             if isinstance(c, str):
+                print("\033[2m%s\033[0m" % c)
                 os.system(c)
             else:
                 c()