+++ CHANGELOG.md
@@ -7,5 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0] - 2020-12
+## [0.0.0] - 2020-12-27
 ### Added
+- This is the first (unreleased) version of Malef.
+- Almost all types has been declared and defined.
+- The resize event has been added.
+- The controlled Surface_Type has been added.
+- A wrapper function has been added.
+- System specific implementations has been added for both Windows and Linux.
+